refactor: split compaction module
Split the compaction monolith into focused modules for hints, instruction building, payload shaping, and trigger orchestration while keeping the public completion API stable.
